texture color code purification

+++ includes/cub_image.h
@@ -60,7 +60,7 @@ typedef struct	s_tex_col
 {
 	int			tex_x;
 	int 		tex_y;
-	int 		tex_step;
+	double 		tex_step;
 	int			wall_side;
 	double 		tex_pos;
 }				t_tex_col;
